+++ src/main/java/learn/springang/umsportal/repositories/UserRepository.java
@@ -1,0 +1,13 @@
+package learn.springang.umsportal.repositories;
+
+import learn.springang.umsportal.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    Optional<User> findByUsername(String username);
+
+    Optional<User> findByEmail(String email);
+}