+++ src/main/java/learn/springang/umsportal/services/UserService.java
@@ -1,0 +1,4 @@
+package learn.springang.umsportal.services;
+
+public interface UserService {
+}